+++ operators/coord.py
@@ -28,11 +28,3 @@ class ViewportCoord(bpy.types.Operator):
         self.list_location_2d[:] = []
 
         return {'FINISHED'}
-
-# [Vector((1302.6494140625, 466.4871826171875)), Vector((1265.6571044921875, 466.4871826171875)), Vector((1300.076904296875, 470.3946838378906)), Vector((1297.2320556640625, 473.6871643066406)), Vector((1294.163330078125, 476.3083190917969)), Vector((1290.92333984375, 478.2132568359375)), Vector((1287.5675048828125, 479.3694152832031)), Vector((1284.1533203125, 479.75701904296875)), Vector((1280.739013671875, 479.3694152832031)), Vector((1277.3831787109375, 478.2132568359375)), Vector((1274.1431884765625, 476.3083190917969)), Vector((1271.0745849609375, 473.6871643066406)), Vector((1268.2296142578125, 470.3946838378906)), Vector((1302.5997314453125, 463.6097412109375)), Vector((1302.7255859375, 460.36724853515625)), Vector((1303.2095947265625, 456.9804382324219)), Vector((1304.0435791015625, 453.5072326660156)), Vector((1305.213134765625, 450.007080078125)), Vector((1306.6982421875, 446.53985595703125)), Vector((1308.4736328125, 443.1648864746094)), Vector((1265.704833984375, 463.61102294921875)), Vector((1265.5767822265625, 460.3684387207031)), Vector((1265.0906982421875, 456.98138427734375)), Vector((1264.2548828125, 453.5078125)), Vector((1263.0836181640625, 450.0071716308594)), Vector((1261.596923828125, 446.5393371582031)), Vector((1259.8203125, 443.1636657714844)), Vector((1284.1533203125, 429.12353515625))]
-
-
-# [Vector((1261.04345703125, 394.4236145019531)), Vector((1242.54736328125, 426.4598388671875)), Vector((1256.373291015625, 394.6976623535156)), Vector((1252.099365234375, 395.51519775390625)), Vector((1248.2950439453125, 396.8622131347656)), Vector((1245.025390625, 398.71563720703125)), Vector((1242.34619140625, 401.04376220703125)), Vector((1240.3033447265625, 403.8067932128906)), Vector((1238.931884765625, 406.9573974609375)), Vector((1238.2552490234375, 410.44171142578125)), Vector((1238.2850341796875, 414.2001037597656)), Vector((1239.0206298828125, 418.1682434082031)), Vector((1240.4495849609375, 422.27825927734375)), Vector((1263.5106201171875, 395.9053649902344)), Vector((1284.1533203125, 401.0408630371094)), Vector((1280.3427734375, 400.890869140625)), Vector((1276.5975341796875, 400.44342041015625)), Vector((1272.9815673828125, 399.7062072753906)), Vector((1269.556640625, 398.6918029785156)), Vector((1266.381591796875, 397.4175720214844)), Vector((1245.06201171875, 427.85662841796875)), Vector((1247.80615234375, 429.58880615234375)), Vector((1250.496337890625, 431.7032775878906)), Vector((1253.086669921875, 434.1639099121094)), Vector((1255.53271484375, 436.9285888671875)), Vector((1257.7926025390625, 439.949951171875)), Vector((1259.8277587890625, 443.1763916015625)), Vector((1284.1533203125, 429.12353515625))]
-
-
-# [Vector((1307.2630615234375, 394.4236145019531)), Vector((1325.7591552734375, 426.4598388671875)), Vector((1311.9332275390625, 394.6976623535156)), Vector((1316.2071533203125, 395.51519775390625)), Vector((1320.011474609375, 396.8622131347656)), Vector((1323.2811279296875, 398.71563720703125)), Vector((1325.9603271484375, 401.04376220703125)), Vector((1328.003173828125, 403.8067932128906)), Vector((1329.3746337890625, 406.9573974609375)), Vector((1330.05126953125, 410.44171142578125)), Vector((1330.021484375, 414.2001037597656)), Vector((1329.285888671875, 418.1682434082031)), Vector((1327.85693359375, 422.27825927734375)), Vector((1304.7958984375, 395.9053649902344)), Vector((1301.9249267578125, 397.4176025390625)), Vector((1298.7498779296875, 398.6918029785156)), Vector((1295.3250732421875, 399.7062072753906)), Vector((1291.708984375, 400.44342041015625)), Vector((1287.9637451171875, 400.890869140625)), Vector((1284.1533203125, 401.0408630371094)), Vector((1323.2445068359375, 427.85662841796875)), Vector((1320.5003662109375, 429.58880615234375)), Vector((1317.8101806640625, 431.7032775878906)), Vector((1315.2198486328125, 434.1639099121094)), Vector((1312.7738037109375, 436.9285888671875)), Vector((1310.513916015625, 439.949951171875)), Vector((1308.478759765625, 443.1763916015625)), Vector((1284.1533203125, 429.12353515625))]
